requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
outcome==1.2.0
prompt-toolkit==3.0.36
pycparser==2.21
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
redis==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.10.3
tweepy==4.14.0
tzdata==2022.7
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.5
wsproto==1.2.0